+++ source/EngineGpuKernels/Genome.cuh
@@ -2,12 +2,16 @@
 
 #include <cstdint>
 
+#include <cuda_fp16.h>
+
 #include <EngineInterface/CellTypeConstants.h>
 #include <EngineInterface/SimulationParametersTypes.h>
 
+// NeuralNetworkGenome weights stored as half precision (FP16) for memory efficiency
+// and consistency with NeuralNetwork/NeuralNetworkTO/NeuralNetworkGenomeTO
 struct NeuralNetworkGenome
 {
-    float weights[MAX_CHANNELS * MAX_CHANNELS];
+    half weights[MAX_CHANNELS * MAX_CHANNELS];  // FP16 for tensor cores
     float biases[MAX_CHANNELS];
     ActivationFunction activationFunctions[MAX_CHANNELS];
     float connectionWeights[MAX_OBJECT_CONNECTIONS];
